refactor(test): reorganize test structure from flat to module-based hierarchy
- Remove legacy test files in tests/modules/ for fantia and todo modules
- Add new structured test under tests/unit/modules/fantia/ following project conventions
- This consolidates test organization and improves maintainability
